requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
dataclasses-json==0.6.2
distro==1.8.0
frozenlist==1.4.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
idna==3.4
jsonpatch==1.33
jsonpointer==2.4
langchain==0.0.334
langsmith==0.0.63
lxml==4.9.3
marshmallow==3.20.1
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.26.1
openai==1.2.3
outcome==1.3.0.post0
packaging==23.2
pandas==2.1.2
Pillow==10.1.0
pydantic==2.4.2
pydantic_core==2.10.1
PyPDF2==3.0.1
PySocks==1.7.1
python-dateutil==2.8.2
python-docx==1.1.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
selenium==4.15.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.23
tenacity==8.2.3
tqdm==4.66.1
trio==0.23.1
trio-websocket==0.11.1
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
wsproto==1.2.0
yarl==1.9.2
